backend/requirements.txt
uvicorn~=0.30.6
fastapi~=0.112.2
pydantic_settings~=2.4.0
tortoise-orm~=0.21.6
loguru~=0.7.2
PyJWT~=2.9.0
PyYAML~=6.0
bcrypt~=4.2.0
aiosmtplib~=3.0.2
python-multipart~=0.0.9
aiofiles~=24.1.0
